+++ src/senaite/core/upgrade/v02_06_000.py
@@ -2480,3 +2480,13 @@ def fix_corrupted_transitions(tool):
                 wapi.update_transition(transition, after_script="")
 
     logger.info("Fixing corrupted transitions [DONE]")
+
+
+def reindex_analysis_categories(tool):
+    logger.info("Reindexing analysis categories ...")
+    cat = api.get_tool(SETUP_CATALOG)
+    for brain in cat(portal_type="AnalysisCategory"):
+        obj = brain.getObject()
+        logger.info("Reindex analysis category: %r" % obj)
+        obj.reindexObject(idxs=["sortable_title"], update_metadata=False)
+    logger.info("Reindexing analysis categories [DONE]")